Backend Services (6):   1. redisSessionService.ts - Session persistence   2. discountCodeService.ts - Discount validation   3. addressVerificationService.ts - Address checking   4. paymentTimeout.ts - Timeout utilities   5. checkoutRateLimiter.ts - Rate limiting   6. checkoutRoutes.enhanced.ts - Updated routes
  Frontend (1):
  7. checkoutReducer.ts - State management

  Tests (4):
  8. checkout.test.ts - Core tests (31 tests)
  9. checkoutController.test.ts - Integration (29 tests)
  10. checkoutSecurity.test.ts - Security (40 tests)
  11. checkoutReducer.test.ts - Frontend (39 tests)

+++ app/contracts/contracts/ReputationSystem.sol
@@ -350,7 +350,7 @@ contract ReputationSystem is ReentrancyGuard, Ownable {
         ReputationScore storage score = reputationScores[user];
         
         // Calculate new average rating
-        uint256 totalRating = score.averageRating * score.reviewCount + (rating * 100);
+        uint256 totalRating = score.averageRating * score.reviewCount + (uint256(rating) * 100);
         score.reviewCount++;
         score.averageRating = totalRating / score.reviewCount;
